// src/app/infraestructure/dto/response/tablaReferencial.dto.ts
export interface TableReferencialResponseDto {
    tableId: number;
    languageId: number;
    codeId: number;
    textCode: string;
    description1: string;
    description2: string;
    auxiliaryData1: string;
    auxiliaryData2: string;
    auxiliaryData3: string;
    parameter: string;
    active: boolean;
    dateRegister: string;
    dateUpdate: string;
    personIdCreate: number;
    personIdUpdate: number;
  }
  